Make systemimage builder generic; build pinetab
Make it possible to build more system-image images by making the builder
more generic, allowing it to take image names as parameters. Then use
that to build a PineTab image too.

Also create the ext4 partitions with the metadata_csum feature disabled
so the 16.04 e2fsck will work on them.