Do not dig holes
For some reason, this --dig-holes causes images created by debos to not
work correctly when flashed with bmaptool over an unclean SD card.
Removing hole digging makes things better.

Unfortunately this adds about 1GB to flash time again and most of it is
zeroes. But at least it doesn't cause the XZ size to balloon again.